requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# src/twitchanal/secret/secret.py: 2
PyYAML == 5.3.1

# src/twitchanal/collect/save.py: 12
alive_progress == 1.6.1

# src/twitchanal/collect/fetch.py: 9
beautifulsoup4 == 4.9.3

# src/twitchanal/cli.py: 1
click == 7.1.2

# src/twitchanal/collect/fetch.py: 1
# src/twitchanal/collect/save.py: 3
pandas == 1.1.4

# src/twitchanal/collect/fetch.py: 2
requests == 2.24.0

# src/twitchanal/collect/fetch.py: 7
# src/twitchanal/secret/secret.py: 3
termcolor == 1.1.0

# src/twitchanal/collect/fetch.py: 8
# src/twitchanal/collect/save.py: 6,7,8
twitchAPI == 2.0.1
